Convert classes to factories
Object oriented programming isn't native to LUA, so instead of forcing
it to behave the same way it seems better to approach it in a more lua
specific way. When using the WoW function CreateFrame, a unique
instance is already being created. Functionality can be attached to it
without having to create a separate "Icon" instance, for example.

"Icon" is a Frame, and the Icon factory can add the variables/functions
that make Icon unique from other Frames.