Change Evaluators data structure
Previously, the data structure was:
Evaluator.Comparison.ComparisonValue.listeners. Now, Evaluate is
expected to return a key and evaluatedState. The key, set in
GetValueGroup, is calculated by concatenating listener.inputValues.
This change makes it necessary for inputValues to only have string and
number values, rather than being key/value pairs.

These changes simplify the code, but also for listeners to have as many
or as few inputValues as necessary.