Allow ValueGroups to save values for use by the evaluator
Some evaluators will need to save values per value group, which they
use to calculate the overall state of the evaluator. An example is
detecting whether an instant cast spell was cast. The only way to do
this is to keep track of whether the unit was casting before the
UNIT_SPELLCAST_SUCCEEDED event goes off.