Create CombatLogEventValueNames.lua file
This table is to be used by Combat_Log to reformat its data so that
it's easily searchable by other evaluators. There's data that is common
to events with the same suffix (such as AURA_BROKEN always having the
value "auraType"), but the location of it in the table will depend on
the prefix.