Replace savedValues with saving directly to ValueGroup
Since the ValueGroup saves a reference to its inputValues, there's
value in referncing it direclty instead of looking up the infromation
from other sources in Evaluate. If Evaluate has been called, it can be
assumed that the ValueGroup's inputValues are a match, since GetKey
gets evaluated first.

Since ValueGroup is being passed to Evaluate, there's no need to have
a separate concept of savedValues. Values can be saved directly to the
ValueGroup, removing the need to create a savedValues table.