Change default Evaluator inputValue for empty strings
By having the default value as "_" instead of "", it allows for
individual listeners to decide if a specific inputValue is important
to them.

In the example of Combat_Log, there may be cases where the listener
wants to know if the destUnit is "player", but doesn't care about the
sourceUnit. Before, the key would be "AURA_NAMEdestUnit". The
problem this creates is that destUnit is likely a valid sourceUnit
value, so what should be two different ValueGroups get put together.

As
an example:

ValueGroup A
[[sourceUnit]] = player
[[destUnit]] = target
[[eventName]] = SPELL_AURA_APPLIED

ValueGroup B
[[sourceUnit]] = player
[[destUnit]] = ""
[[eventName]] = SPELL_AURA_APPLIED

ValueGroup C
[[sourceUnit]] = ""
[[destUnit]] = player
[[eventName]] = SPELL_AURA_APPLIED

ValueGroup A key = playertargetSPELL_AURA_APPLIED
ValueGroup B key = playerSPELL_AURA_APPLIED
ValueGroup C key = playerSPELL_AURA_APPLIED