Change how EVENT type evaluators pass values
The core change is that EVENT type evaluators will not generally send
events themselves. Instead, they'll send data directly to their
listeners. Without this, listeners would be expected to parse event
data themselves.

As an example, an evaluator that checks if the UIObject that had its
visibility changed has specific tags. After it performs this
evaluation, there will be specific listeners that need to be informed
that a UI object with a specific set of tags was shown or hidden. It
would be possible to send a concatenated version of the inputValues,
but then the listeners would have to concatenate their own inputValues
and compare. This would be a core breakdown of the relationship
between evaluators and listeners.