Separate ListenerGroups value handling into ValueHandlers
What prompted this overhaul was working with the concept of sorting
children. It's possilbe that elements would be sorted by a preset value
that never changes. Having ListenerGroups act as static value holders
is out of scope for that data structure. Separating value handling into
its own well defined data structure is much cleaner and flexible.

The functions in SortGroups, StateChange, and Teardown and most of
the functions in Setup have already been moved to ListenerGroups or
ValueHandlers.

* ValueHandlers must have a key.
* ValueHandlers must assign a function to ChangeValue.
* ValueHandlers can assign a function to Setup.
* StateValue type ValueHandlers must assign:
    - Function to OnValidValue
    - Function to OnInvalidValue
    - Table of integer key/value pairs to validValues.
* All ListenerGroups must assign a function to OnEvaluate.
* State type ListenerGroups must assign a string to valueHandlerKey.