Ensure that frames sizes are never set to zero
Groups that are dynamically sized won't have a starting width or
height. If they have a width or height of 0, their child frames will
never appear. When the group goes to resize itself based on the size of
its child frames, it will find that all of their sizes are 0.
Essentially, nothing will appear on the screen.