Refactor EvaluatorsHelpers function naming and structure
Naming changes are to increase readability and make the context of
functions obvious right at the start to differentiate similar functions.

Split Evaluate into EvaluateAsState and EvaluateAsEvent for
readability. Evaluators need to have a more thorough refactoring pass
as it feels like State and Event evaluators are different enough to
be different types, but should share the same base type.