Change Components file structure and namespace
The intent is to differentiate between full-function components and
elements used to create components. There are only three components
as of right now: Children, EnabledState, and VisibleState.

This also reflected that Components had been moved out from under
the UI.Objects hierarchy and placed as a direct child of UI.