Convert UI\Objects files to use component and template systems
This greatly cuts down on duplicated code as the only values that
should be assigned are those that are unique to the UIObject in
question. The required values can be found at the top of UI\Components
files, though it's important not to forget to follow the entire
inheritence chain to ensure all required values are assigned.