Refactor Listeners
The correct functionality of ValueChanger is that it should take a
value assigned to a Listener and do something with it.

Along with this,
all Listeners should be EvaluatorListeners. Only EventEvaluators should
be adding themselves as listeners of Events directly. As such,
Registering/Unregistering to Evaluators is now handled by
Listeners.Base.