Remove unnecessary nil value check for UIObject
If a listener has registered a specific UIObject with this event
evaluator, it should be assumed that the UIObject exists and that the
required component exists as well. If it doesn't, an error needs to be
thrown since that means either the listener shouldn't be registering
this UIObject, or adding the required component to the UIObject has
been forgotten.