Change when Parent DisplayUpdates are fired
The UIOBJECT_WITH_PARENT_VISIBLE_CHANGED was being called before
UIOBJECT_VISIBLE_CHANGED, which is what the Child component bases its
updates off of. Because of this, Parent's DisplayUpdate was firing
prior to children being registered, causing the arrangement to be based
on only the children that were registered to it prior to the new
becoming visible.

Since children are calling ChildRegister when they become visible, the
DisplayUpdate calls triggered by their visibility state changing should
be handled by the functions the children call when their visibility
state changes (ChildRegister and ChildDeregister).