Don't fire event when using SetDimensions
This function is only used by Frame and Pools, both of which are
involved in setting up frames for use. This causes duplicate
DisplayUpdate calls every time a child becomes visible, since it's
being called when the (1) the child frame is created and (2) when the
child frame is registered with Parent.