Correctly set minimum size in SetDimensions
A width or height of 0 causes the frame's children to have a width and
height value of nil. In cases where there's only one icon--as is common
with the Active module--the group's size won't be correct until two
icons appear. By setting a correct min value initially, this won't
happen.