Refactor how Frame.SetDimensions sets the parentFrame value
By having parentFrame as an argument, it forces Frame.Create and
Pools:Claim to pass UIParent so the value won't be nil. In creating
Cooldown frames, it's important that the parent is declared prior to a
call to Pools:Claim so that SetDimensions doesn't have to be called a
second time after the frame is created.