Refactor components to use UIOBJECT_COMPONENT_STATE_CHANGED
A core side effect of this change is a change in the relationship
between VisibleState and Frame. The concept of frame "users" is gone,
replaced by frame "modifiers." Instead of VisibleState being a user of
Frame, it's now how Frame determines its own state.

Event evaluators that were dedicated to the state of a specific
component are now redundant and have been deleted.