Activate elements during setup instead of OnEnable
Having elements activate in reaction to the UIObject's EnableState
component state caused issues with listeners being called in the
correct order.

As an example, the MindFlay UIObject was registering itself as a child
of the PrimaryModule UIObject before PrimaryModule had a frame. This
caused the module to not appear until after the MindFlay icon
disappeared and reappeared.

It's now possible for UIOBJECT_COMPONENT_STATE_CHANGED to be called to
check the state of a component before the component exists, so a nil
check has been added to account for this situation.