Check that a component instance hasn't been setup
The Frame base component assigns a "Frame" value to the UIObject,
which is a pointer to the component that derives from Frame. The
UIObjectSetup function will treat this pointer as though it's a
component itself, calling the Setup function. This causes duplicate
listeners to be registered, resulting in some UIObjects registering as
children of a Group twice (Child.OnShow gets called twice).