Update charge count when not cast recently
Previously, the timer to update charge counts would only be triggered
if the evaluation was triggered by the player successfully casting a
spell. This does not account for cases where the player is allowing
charges to increase without recasting the spell.

Example: a spell has a maximum of 2 charges. The player casts the spell
until it has 0 charges. Once the charge cooldown is complete, the
charge count increases to 1. The player casts other spells instead of
this spell again, allowing the charge count to reach 2. This was not
being reevaluated since TimerStart was only being called on cast
success events.