Delay update of spell charge count by a single frame
The charge count for spells doesn't update until the frame after
SPELL_CAST_SUCCESS is called, and there are other event evaluators that
also have to delay their update by a frame. By updating this event
evaluator a frame before the others, it causes icons to flicker for a
single frame.