Update GAMEDATA_LOADED state when player leaves world
When the player moves from one instance to another,
PLAYER_ENTERING_WORLD is fired again on them reentering the world. This
causes the IntegerKeyState value of UIParentGroup's EnabledState
listener to be increased. Since PLAYER_LEAVING_WORLD wasn't being
listener for, this value was never decreasing. The end result is that
the key value was only correct on game reload.