Refactor frame modification calls from FrameModifier to Frame
This change allows FrameModifiers to control whether or not they should
modify a frame, rather than having to do that and control when the
frame should be modified. The Frame base component is the best place to
handle the timing of frame modification as it can guarantee a frame
will be available for modification.