Create UNIT_COUNT_CLOSE_TO_UNIT_CHANGED event evaluator
This evaluator replaces UNIT_COUNT_IN_RANGE_OF_UNIT_CHANGED as there's
no way to get the actual range of spells from game data. Maintaining a
list of spell radiuses isn't an ideal option. The compromise is not
worrying about the actual range by limiting the scope to the concept of
"close to."

The core thought process is that units who are damaged by the same AoE
spell are likely "close to" each other. Units that are damaged by
multiple AoE spells in a short period of time are pretty much
guaranteed to be to be "close to" each other. By parsing combat event
logs, it's possible to track what units were damaged by AoE spells.

A further consideration is that it needs to be possible to check the
number of units around a friendly unit. See the TODO comment above
Evaluate for further thoughts on this. This functionality is only
partially implemented.