Remove spell events faster than melee events
When a melee unit is running, it's possible that enemies chasing it
won't be able to get any melee swings in for a few seconds. Other range
units may be attacking the enemies with AoE spells cast on the melee
unit (such as Mind Sear), so it's important that those enemies still be
listed as valid melee units. With the melee events having the same max
elapsed time, they fall off too soon and cause the AoE spells to stop
registering as valid.