Delay declaiming of frames
This is to avoid situations where a frame was released and then claimed
during the same event. This causes the SetPoints call in ChildArrangers
to throw an error due to HUD_GROUP releasing its frame, then
HUD_MODULE_SECONDARY trying to arrange a child that's using the frame
HUD_GROUP released earlier in the event.