Increase floor comparison value
Due to the "Between" comparison now being inclusive, a value of 0 would
trigger a true state for CastSoonAlert during the first cast. This is
due to PLAYER_SPELL_COOLDOWN_DURATION_CHANGED using 0 to represent the
cooldown being a GCD during setup. The chance of
PLAYER_SPELL_COOLDOWN_DURATION_CHANGED returning exactly 0 outside of
it being a GCD is very low, but it's possible for the difference to be
close to 0 (such as 0.1).