Refactor how icon displayers are managed
Having a single displayer value per icon causes problems when a
displayer's state goes from false to true to false in a single frame.
If it had a displayer prior to this it would get wiped in the process.
Keeping individual states for each displayer prevents displayer values
from being lost.