Reduce chance of initial value incorrectly being 0
When a player first attacks an enemy, their damage per second while
attacking that enemy will look very high even if they did very little
damage. This is because the time elapsed since the first interaction
with that target (what lookbackDuration is set to towards the start of
the fight) is very short. As an example, doing 100 damage in 0.01
seconds is 10,000 dps. As such, we need to provide a minimum
lookbackDuration to ensure that a more reasonable first value is
returned.